requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
Brotli==1.0.9
colorama==0.4.6
decorator==5.1.1
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.3
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
model-bakery==1.15.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pycodestyle==2.11.0
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-testdox==3.1.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.4
wcwidth==0.2.6
whitenoise==6.5.0
